Major rewrite and performance optimization
This is a major update. In addition to addressing issues raised by the
team, the code has had a major overhaul for style, structure and
performance.
Execution speed has more than doubled and the code has had a major
overhaul. It has been tested and the output is identical to the previous
version.
This update addresses the following issues (closed with appropriate
comments):
#13 Hash table of floats increasing by 0.5?
#12 Is catching an exception to rethrow really adding value here?
#11 Hand-crafted spin loops for thread completion?
#10 com.java package
#8 How many times do I have to parse the same number from a string?
#7 ArrayList appears to exist solely to have its minimum taken
#6 Indentation
#5 Are you using a HashMap on sequential integers keys?